Optimise TSP
Added a better lower bound calculation and an iterator using TreeMap to go through the list faster

+++ com.H4143.PLD-AGILE/src/main/java/tsp/TemplateTSP.java
@@ -173,7 +173,7 @@ public abstract class TemplateTSP implements TSP {
 	 * @return a lower bound of the cost of paths in <code>g</code> starting from <code>currentVertex</code>, visiting 
 	 * every vertex in <code>unvisited</code> exactly once, and returning back to vertex <code>0</code>.
 	 */
-	protected abstract int bound(Integer currentVertex, Collection<Integer> unvisited);
+	protected abstract double bound(Integer currentVertex, Collection<Integer> unvisited);
 	
 	/**
 	 * Method that must be defined in TemplateTSP subclasses
@@ -208,7 +208,7 @@ public abstract class TemplateTSP implements TSP {
 	        	visited.add(nextVertex);
 	            unvisited.remove(nextVertex);
 	            if (nextVertex%2==1) unvisited.add(nextVertex+1);
-	            branchAndBound(nextVertex, unvisited, visited, 
+	            branchAndBound(nextVertex, unvisited, visited,  
 	            		currentCost+g.getCost(currentVertex, nextVertex));
 	            visited.remove(nextVertex);
 	            unvisited.add(nextVertex);
